+++ src/main/java/com/example/demo/service/reservation/ReservationService.java
@@ -1,7 +1,7 @@
-package com.example.demo.service;
+package com.example.demo.service.reservation;
 
 import com.example.demo.entity.Reservation;
-import com.example.demo.util.ReservationRequest;
+import com.example.demo.util.reservation_request.ReservationRequest;
 
 public interface ReservationService {
     Reservation processReservationRequest(ReservationRequest reservationRequest);
