+++ src/main/java/com/example/demo/service/room/RoomService.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.service.room;
 
 import com.example.demo.entity.Room;
