+++ src/main/java/com/example/demo/service/seat/SeatService.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.service.seat;
 
 import com.example.demo.entity.Seat;
