+++ src/main/java/com/example/demo/service/ticket_price/TicketPriceService.java
@@ -1,6 +1,6 @@
-package com.example.demo.service;
+package com.example.demo.service.ticket_price;
 
-import com.example.demo.util.TicketType;
+import com.example.demo.util.reservation_request.TicketType;
 
 import java.util.Map;
