+++ src/main/java/com/example/demo/util/reservation_request/CustomerData.java
@@ -1,4 +1,4 @@
-package com.example.demo.util;
+package com.example.demo.util.reservation_request;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
