+++ src/main/java/com/example/demo/util/reservation_request/TicketType.java
@@ -1,4 +1,4 @@
-package com.example.demo.util;
+package com.example.demo.util.reservation_request;
 
 public enum TicketType {
     ADULT,
